Add files via upload
A Smart Device for Cloth Protection from Rain is an innovative system designed to safeguard clothes left outdoors during unexpected rain. It uses a rain sensor to detect rainfall and an Arduino Uno to control the mechanism. Upon detecting rain, the system triggers an IC555 timer to manage the timing of the operation, activating a servo motor to pull a protective cover over the clothes via a rope mechanism. Additionally, a temperature sensor is integrated to ensure efficient functionality in varying weather conditions. This automated system provides a practical, energy-efficient solution for protecting clothes from rain.